Make casts to unsigned int explicit.

+++ src/doom/r_plane.cpp
@@ -168,7 +168,7 @@ R_MapPlane
 	ds_colormap[0] = ds_colormap[1] = fixedcolormap;
     else
     {
-	index = distance >> LIGHTZSHIFT;
+       index = (unsigned int)(distance >> LIGHTZSHIFT);
 	
 	if (index >= MAXLIGHTZ )
 	    index = MAXLIGHTZ-1;
